+++ zcash_primitives/src/transaction/components/sapling/builder.rs
@@ -272,14 +272,14 @@ impl<P: consensus::Parameters> SaplingBuilder<P> {
         merkle_path: MerklePath<Node>,
     ) -> Result<(), Error> {
         // Consistency check: all anchors must equal the first one
-        let cmu = Node::new(note.cmu().into());
+        let node = note.commitment();
         if let Some(anchor) = self.anchor {
-            let path_root: bls12_381::Scalar = merkle_path.root(cmu).into();
+            let path_root: bls12_381::Scalar = merkle_path.root(node).into();
             if path_root != anchor {
                 return Err(Error::AnchorMismatch);
             }
         } else {
-            self.anchor = Some(merkle_path.root(cmu).into())
+            self.anchor = Some(merkle_path.root(node).into())
         }
 
         let alpha = jubjub::Fr::random(&mut rng);